@@ -1,14 +1,14 @@
 """This file holds the BeeAI backend implementation."""
 
-import json
 import datetime
-from typing import Any, Dict, List, Optional, Type
+import json
+from typing import Any, Optional
 
 from beeai_framework.backend import Backend as BeeAIFrameworkBackend
 from beeai_framework.backend.chat import ChatModel
 from beeai_framework.backend.constants import ProviderName
+from beeai_framework.backend.message import AssistantMessage, SystemMessage, UserMessage
 from beeai_framework.backend.types import ChatModelParameters
-from beeai_framework.backend.message import SystemMessage, UserMessage, AssistantMessage
 
 from mellea.backends import BaseModelSubclass
 from mellea.backends.formatter import FormatterBackend
@@ -31,12 +31,11 @@ class BeeAIBackend(FormatterBackend):
     def __init__(
         self,
         model_id: str,
-        formatter,
-        *,
-        api_key: Optional[str] = None,
-        base_url: Optional[str] = None,
-        provider: Optional[str] = None,
-        model_options: Optional[Dict[str, Any]] = None,
+        formatter: Any | None = None,
+        provider: str | None = None,
+        base_url: str | None = None,
+        api_key: str | None = None,
+        model_options: dict[str, Any] | None = None,
     ):
         """Initialize the BeeAI backend.
 
@@ -48,20 +47,22 @@ class BeeAIBackend(FormatterBackend):
             provider: The provider to use (e.g., 'openai', 'anthropic', 'local').
             model_options: Default model options for this backend.
         """
-        super().__init__(model_id=model_id, formatter=formatter, model_options=model_options or {})
-        
+        super().__init__(
+            model_id=model_id, formatter=formatter, model_options=model_options or {}
+        )
+
         self.api_key = api_key
         self.base_url = base_url
-        
+
         # Auto-detect provider based on model_id for local Ollama models
-        if model_id.startswith("granite") or model_id.startswith("llama") or model_id.startswith("mistral"):
+        if model_id.startswith(("granite", "llama", "mistral")):
             self.provider = "ollama"
             # Set default Ollama base URL if not provided
             if not self.base_url:
                 self.base_url = "http://localhost:11434"
         else:
             self.provider = provider or "openai"  # Default to OpenAI-compatible
-        
+
         # Initialize the BeeAI backend
         try:
             # Create a chat model from the provider name
@@ -72,7 +73,7 @@ class BeeAIBackend(FormatterBackend):
                 chat_kwargs["api_key"] = self.api_key
             if self.base_url:
                 chat_kwargs["base_url"] = self.base_url
-            
+
             # For testing purposes, we'll create the chat model lazily
             self._chat_kwargs = chat_kwargs
             self._chat_model = None
@@ -86,28 +87,31 @@ class BeeAIBackend(FormatterBackend):
             self._chat_kwargs = chat_kwargs
             self._chat_model = None
             self._beeai_backend = None
-        
+
         # Set up environment variables if API key is provided
         if self.api_key:
             import os
+
             if self.provider == "openai":
                 os.environ["OPENAI_API_KEY"] = self.api_key
             elif self.provider == "anthropic":
                 os.environ["ANTHROPIC_API_KEY"] = self.api_key
             # Add other providers as needed
-        
+
         if self.base_url:
             if self.provider == "openai":
                 os.environ["OPENAI_BASE_URL"] = self.base_url
             # Add other providers as needed
 
-    def _convert_model_options(self, model_options: Optional[Dict[str, Any]]) -> ChatModelParameters:
+    def _convert_model_options(
+        self, model_options: dict[str, Any] | None
+    ) -> ChatModelParameters:
         """Convert Mellea model options to BeeAI parameters."""
         if not model_options:
             return ChatModelParameters()
-        
+
         beeai_params = {}
-        
+
         # Map Mellea options to BeeAI parameters
         option_mapping = {
             ModelOption.TEMPERATURE: "temperature",
@@ -118,27 +122,27 @@ class BeeAIBackend(FormatterBackend):
             ModelOption.FREQUENCY_PENALTY: "frequency_penalty",
             ModelOption.PRESENCE_PENALTY: "presence_penalty",
         }
-        
+
         for mellea_key, beeai_key in option_mapping.items():
             if mellea_key in model_options:
                 beeai_params[beeai_key] = model_options[mellea_key]
-        
+
         # Handle special cases
         if ModelOption.STOP_SEQUENCES in model_options:
             beeai_params["stop_sequences"] = model_options[ModelOption.STOP_SEQUENCES]
-        
+
         return ChatModelParameters(**beeai_params)
 
     def _convert_context_to_messages(
-        self, 
-        action: Component | CBlock, 
-        ctx: Context
-    ) -> List[Any]:
+        self, action: Component | CBlock, ctx: Context
+    ) -> list[Any]:
         """Convert Mellea context to BeeAI messages."""
         messages = []
-        
+
         # Convert context to messages
-        if hasattr(ctx, 'render_for_generation') and callable(ctx.render_for_generation):
+        if hasattr(ctx, "render_for_generation") and callable(
+            ctx.render_for_generation
+        ):
             linearized_ctx = ctx.render_for_generation()
             if linearized_ctx:
                 for component in linearized_ctx[:-1]:  # All but the last (action)
@@ -153,7 +157,7 @@ class BeeAIBackend(FormatterBackend):
                         # Convert component to string using formatter
                         content = self.formatter.print(component)
                         messages.append(UserMessage(content=content))
-        
+
         # Add the action as the last message
         if isinstance(action, Message):
             if action.role == "system":
@@ -165,7 +169,7 @@ class BeeAIBackend(FormatterBackend):
         else:
             content = self.formatter.print(action)
             messages.append(UserMessage(content=content))
-        
+
         return messages
 
     def generate_from_context(
@@ -173,33 +177,36 @@ class BeeAIBackend(FormatterBackend):
         action: Component | CBlock,
         ctx: Context,
         *,
-        format: Optional[Type[BaseModelSubclass]] = None,
-        model_options: Optional[Dict[str, Any]] = None,
-        generate_logs: Optional[List[GenerateLog]] = None,
+        format: type[BaseModelSubclass] | None = None,
+        model_options: dict[str, Any] | None = None,
+        generate_logs: list[GenerateLog] | None = None,
         tool_calls: bool = False,
     ) -> ModelOutputThunk:
         """Generate a response from context using the BeeAI backend."""
         try:
             # Convert context to BeeAI messages
             messages = self._convert_context_to_messages(action, ctx)
-            
+
             # Convert model options
             beeai_params = self._convert_model_options(model_options)
-            
+
             # Generate response
             chat_model = self._get_chat_model()
             response = chat_model.create(
-                messages=messages,
-                **beeai_params.model_dump(exclude_none=True)
+                messages=messages, **beeai_params.model_dump(exclude_none=True)
             )
-            
+
             # Extract the response content from ChatModelOutput
             # For testing purposes, if response is a Mock, extract the content directly
-            if hasattr(response, '_mock_name') or hasattr(response, '_mock_return_value') or str(type(response)).find('Mock') != -1:
+            if (
+                hasattr(response, "_mock_name")
+                or hasattr(response, "_mock_return_value")
+                or str(type(response)).find("Mock") != -1
+            ):
                 # This is a mock object, extract content from the mock structure
-                if hasattr(response, 'messages') and response.messages:
+                if hasattr(response, "messages") and response.messages:
                     first_message = response.messages[0]
-                    if hasattr(first_message, 'content'):
+                    if hasattr(first_message, "content"):
                         content = str(first_message.content)
                     else:
                         content = str(first_message)
@@ -207,32 +214,34 @@ class BeeAIBackend(FormatterBackend):
                     content = str(response)
             else:
                 # Real BeeAI response - extract from ChatModelOutput
-                if hasattr(response, 'messages') and response.messages:
+                if hasattr(response, "messages") and response.messages:
                     # Get the content from the first message
                     first_message = response.messages[0]
-                    if hasattr(first_message, 'content'):
+                    if hasattr(first_message, "content"):
                         if isinstance(first_message.content, list):
                             # Content might be a list of content blocks
-                            content = " ".join([str(block) for block in first_message.content])
+                            content = " ".join(
+                                [str(block) for block in first_message.content]
+                            )
                         else:
                             content = str(first_message.content)
                     else:
                         content = str(first_message)
-                elif hasattr(response, 'content') and response.content:
+                elif hasattr(response, "content") and response.content:
                     content = response.content
-                elif hasattr(response, 'message') and response.message:
+                elif hasattr(response, "message") and response.message:
                     content = response.message.content
-                elif hasattr(response, 'text') and response.text:
+                elif hasattr(response, "text") and response.text:
                     content = response.text
                 else:
                     content = str(response)
-            
+
             # Create ModelOutputThunk
             result = ModelOutputThunk(value=content)
-            
+
             # Parse the result using the formatter
             self.formatter.parse(action, result)
-            
+
             # Log generation if requested
             if generate_logs is not None:
                 generate_log = GenerateLog()
@@ -244,47 +253,50 @@ class BeeAIBackend(FormatterBackend):
                 generate_log.action = action
                 generate_log.result = result
                 generate_logs.append(generate_log)
-            
+
             return result
-            
+
         except Exception as e:
             FancyLogger.get_logger().error(f"BeeAI generation failed: {e}")
             raise
 
     def _generate_from_raw(
         self,
-        actions: List[Component | CBlock],
+        actions: list[Component | CBlock],
         *,
-        format: Optional[Type[BaseModelSubclass]] = None,
-        model_options: Optional[Dict[str, Any]] = None,
-        generate_logs: Optional[List[GenerateLog]] = None,
-    ) -> List[ModelOutputThunk]:
+        format: type[BaseModelSubclass] | None = None,
+        model_options: dict[str, Any] | None = None,
+        generate_logs: list[GenerateLog] | None = None,
+    ) -> list[ModelOutputThunk]:
         """Generate responses from raw actions without context."""
         results = []
-        
+
         try:
             # Convert model options
             beeai_params = self._convert_model_options(model_options)
-            
+
             for action in actions:
                 # Convert action to message
                 content = self.formatter.print(action)
                 message = UserMessage(content=content)
-                
+
                 # Generate response
                 chat_model = self._get_chat_model()
                 response = chat_model.create(
-                    messages=[message],
-                    **beeai_params.model_dump(exclude_none=True)
+                    messages=[message], **beeai_params.model_dump(exclude_none=True)
                 )
-                
+
                 # Extract the response content from ChatModelOutput
                 # For testing purposes, if response is a Mock, extract the content directly
-                if hasattr(response, '_mock_name') or hasattr(response, '_mock_return_value') or str(type(response)).find('Mock') != -1:
+                if (
+                    hasattr(response, "_mock_name")
+                    or hasattr(response, "_mock_return_value")
+                    or str(type(response)).find("Mock") != -1
+                ):
                     # This is a mock object, extract content from the mock structure
-                    if hasattr(response, 'messages') and response.messages:
+                    if hasattr(response, "messages") and response.messages:
                         first_message = response.messages[0]
-                        if hasattr(first_message, 'content'):
+                        if hasattr(first_message, "content"):
                             content = str(first_message.content)
                         else:
                             content = str(first_message)
@@ -292,34 +304,36 @@ class BeeAIBackend(FormatterBackend):
                         content = str(response)
                 else:
                     # Real BeeAI response - extract from ChatModelOutput
-                    if hasattr(response, 'messages') and response.messages:
+                    if hasattr(response, "messages") and response.messages:
                         # Get the content from the first message
                         first_message = response.messages[0]
-                        if hasattr(first_message, 'content'):
+                        if hasattr(first_message, "content"):
                             if isinstance(first_message.content, list):
                                 # Content might be a list of content blocks
-                                content = " ".join([str(block) for block in first_message.content])
+                                content = " ".join(
+                                    [str(block) for block in first_message.content]
+                                )
                             else:
                                 content = str(first_message.content)
                         else:
                             content = str(first_message)
-                    elif hasattr(response, 'content') and response.content:
+                    elif hasattr(response, "content") and response.content:
                         content = response.content
-                    elif hasattr(response, 'message') and response.message:
+                    elif hasattr(response, "message") and response.message:
                         content = response.message.content
-                    elif hasattr(response, 'text') and response.text:
+                    elif hasattr(response, "text") and response.text:
                         content = str(response.text)
                     else:
                         content = str(response)
-                
+
                 # Create ModelOutputThunk
                 result = ModelOutputThunk(value=content)
-                
+
                 # Parse the result using the formatter
                 self.formatter.parse(action, result)
-                
+
                 results.append(result)
-                
+
                 # Log generation if requested
                 if generate_logs is not None:
                     generate_log = GenerateLog()
@@ -331,9 +345,9 @@ class BeeAIBackend(FormatterBackend):
                     generate_log.action = action
                     generate_log.result = result
                     generate_logs.append(generate_log)
-            
+
             return results
-            
+
         except Exception as e:
             FancyLogger.get_logger().error(f"BeeAI raw generation failed: {e}")
             raise
@@ -345,14 +359,11 @@ class BeeAIBackend(FormatterBackend):
                 if self.provider == "ollama":
                     # For Ollama models, we need to pass the model name
                     self._chat_model = ChatModel.from_name(
-                        "ollama",
-                        model=self.model_id,
-                        **self._chat_kwargs
+                        "ollama", model=self.model_id, **self._chat_kwargs
                     )
                 else:
                     self._chat_model = ChatModel.from_name(
-                        self.provider,
-                        **self._chat_kwargs
+                        self.provider, **self._chat_kwargs
                     )
             except Exception as e:
                 FancyLogger.get_logger().warning(
@@ -360,13 +371,12 @@ class BeeAIBackend(FormatterBackend):
                     "Falling back to OpenAI provider."
                 )
                 self.provider = "openai"
-                self._chat_model = ChatModel.from_name(
-                    "openai",
-                    **self._chat_kwargs
-                )
+                self._chat_model = ChatModel.from_name("openai", **self._chat_kwargs)
         return self._chat_model
 
-    def _merge_model_options(self, call_options: Optional[Dict[str, Any]]) -> Dict[str, Any]:
+    def _merge_model_options(
+        self, call_options: dict[str, Any] | None
+    ) -> dict[str, Any]:
         """Merge default model options with call-specific options."""
         merged = self.model_options.copy()
         if call_options:
@@ -376,5 +386,3 @@ class BeeAIBackend(FormatterBackend):
     def __repr__(self) -> str:
         """String representation of the backend."""
         return f"BeeAIBackend(model_id='{self.model_id}', provider='{self.provider}')"
-
-
